+++ front-end/src/main.rs
@@ -1,3 +1,8 @@
+use std::process::ExitCode;
+use std::sync::atomic::AtomicBool;
+use std::sync::atomic::Ordering;
+
+use log::error;
 use pasfmt::format_with_settings;
 use pasfmt::FormattingSettings;
 use pasfmt_orchestrator::predule::*;
@@ -7,13 +12,30 @@ pasfmt_config!(
     Config
 );
 
-fn main() -> anyhow::Result<()> {
+fn main() -> ExitCode {
     let config = Config::create();
     stderrlog::new()
         .verbosity(config.log_level())
         .init()
         .unwrap();
 
-    let formatting_settings = config.get_config_object::<FormattingSettings>()?;
-    format_with_settings(formatting_settings, config)
+    let had_error = AtomicBool::new(false);
+
+    let err_handler = |e| {
+        had_error.store(true, Ordering::Relaxed);
+        error!("{:?}", e);
+    };
+
+    match config.get_config_object::<FormattingSettings>() {
+        Ok(formatting_settings) => {
+            format_with_settings(formatting_settings, config, err_handler);
+        }
+        Err(e) => err_handler(e),
+    }
+
+    if had_error.into_inner() {
+        ExitCode::FAILURE
+    } else {
+        ExitCode::SUCCESS
+    }
 }