+++ orchestrator/src/formatting_orchestrator.rs
@@ -1,7 +1,6 @@
 use std::borrow::Cow;
 
-use crate::{command_line::FormatMode, file_formatter::FileFormatter};
-use anyhow::{anyhow, bail};
+use crate::{command_line::FormatMode, file_formatter::FileFormatter, predule::ErrHandler};
 use log::LevelFilter;
 
 pub trait FormatterConfiguration {
@@ -11,63 +10,28 @@ pub trait FormatterConfiguration {
     fn mode(&self) -> FormatMode;
 }
 
-fn plural<'a>(val: usize, singular: &'a str, plural: &'a str) -> &'a str {
-    if val == 1 {
-        singular
-    } else {
-        plural
-    }
-}
-
-fn coalesce_err(results: &[anyhow::Result<()>]) -> anyhow::Result<()> {
-    match results.iter().filter(|r| r.is_err()).count() {
-        0 => Ok(()),
-        c => Err(anyhow!("{c} {} with errors", plural(c, "file", "files"))),
-    }
-}
-
-fn check<T: FormatterConfiguration>(
-    config: &T,
-    file_formatter: &FileFormatter,
-) -> anyhow::Result<()> {
-    if config.is_stdin() {
-        if file_formatter.check_stdin().is_err() {
-            bail!("<stdin> is incorrectly formatted");
-        }
-    } else {
-        let fail_count = file_formatter
-            .check_files(&config.get_paths()?)
-            .iter()
-            .filter(|r| r.is_err())
-            .count();
-
-        if fail_count > 0 {
-            bail!(
-                "{fail_count} {} incorrectly formatted",
-                plural(fail_count, "file is", "files are")
-            )
-        }
-    }
-
-    Ok(())
-}
-
 pub struct FormattingOrchestrator;
 impl FormattingOrchestrator {
     pub fn run<T: FormatterConfiguration>(
         file_formatter: FileFormatter,
         config: T,
-    ) -> anyhow::Result<()> {
+        err_handler: impl ErrHandler,
+    ) {
         match config.mode() {
-            FormatMode::Check => check(&config, &file_formatter),
-            FormatMode::Files => coalesce_err(&file_formatter.format_files(&config.get_paths()?)),
-            FormatMode::Stdout => {
-                if config.is_stdin() {
-                    file_formatter.format_stdin_to_stdout()
-                } else {
-                    coalesce_err(&file_formatter.format_files_to_stdout(&config.get_paths()?))
-                }
+            FormatMode::Check if config.is_stdin() => file_formatter.check_stdin(err_handler),
+            FormatMode::Stdout if config.is_stdin() => {
+                file_formatter.format_stdin_to_stdout(err_handler)
             }
-        }
+            mode => match config.get_paths() {
+                Ok(paths) => match mode {
+                    FormatMode::Check => file_formatter.check_files(&paths, err_handler),
+                    FormatMode::Files => file_formatter.format_files(&paths, err_handler),
+                    FormatMode::Stdout => {
+                        file_formatter.format_files_to_stdout(&paths, err_handler)
+                    }
+                },
+                Err(e) => err_handler(e),
+            },
+        };
     }
 }